+++ Plugin.cs
@@ -9,7 +9,7 @@ namespace OttoLens;
 public class OttoLensPlugin : BaseUnityPlugin
 {
     internal const string ModName = "OttoLens";
-    internal const string ModVersion = "1.0.0";
+    internal const string ModVersion = "1.0.1";
     internal const string Author = "potto007";
     internal const string ModGUID = $"{Author}.{ModName}";